transaction refinement

+++ Model/AccountDto.cs
@@ -6,7 +6,7 @@ public class AccountDto
     public Guid UserID { get; set; }
     public String Label { get; set; }
     public String Description { get; set; }
-    public int InitValue { get; set; }
+    public decimal InitValue { get; set; }
     public DateTime CreatedAt { get; set; }
     public DateTime UpdatedAt { get; set; }
 }
